Convert arguments to bool

+++ Source/PeleC.cpp
@@ -2075,10 +2075,10 @@ PeleC::reset_internal_energy(amrex::MultiFab& S_new, int ng)
       amrex::ParallelFor(
         bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
           pc_rst_int_e(
-            i, j, k, sarr, static_cast<int>(captured_allow_small_energy),
-            static_cast<int>(captured_allow_negative_energy),
-            static_cast<int>(captured_dual_energy_update_E_from_e),
-            captured_dual_energy_eta2, captured_verbose);
+            i, j, k, sarr, captured_allow_small_energy,
+            captured_allow_negative_energy,
+            captured_dual_energy_update_E_from_e, captured_dual_energy_eta2,
+            captured_verbose);
         });
     }
   }
